move reusable funtions to helper file
